chore: rename commands file
