+++ packages/types/lib/index.ts
@@ -2,9 +2,9 @@ export * from './command-maps';
 export * from './action';
 export * from './appium-config';
 export * from './capabilities';
+export * from './commands';
 export * from './config';
 export * from './constraints';
-export * from './driver-commands';
 export * from './driver';
 export * from './plugin';
 export * from './http';
